moved functionality to acctest module
