+++ names/names_test.go
@@ -601,191 +601,3 @@ func TestAWSGoV1ClientName(t *testing.T) {
 		})
 	}
 }
-
-func TestIsIsolatedPartition(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		input    string
-		expected bool
-	}{
-		{
-			input:    StandardPartitionID,
-			expected: false,
-		},
-		{
-			input:    ChinaPartitionID,
-			expected: false,
-		},
-		{
-			input:    USGovCloudPartitionID,
-			expected: false,
-		},
-		{
-			input:    ISOPartitionID,
-			expected: true,
-		},
-		{
-			input:    ISOBPartitionID,
-			expected: true,
-		},
-		{
-			input:    ISOEPartitionID,
-			expected: true,
-		},
-		{
-			input:    ISOFPartitionID,
-			expected: true,
-		},
-	}
-
-	for _, testCase := range testCases {
-		testCase := testCase
-		t.Run(testCase.input, func(t *testing.T) {
-			t.Parallel()
-
-			if got, want := IsIsolatedPartition(testCase.input), testCase.expected; got != want {
-				t.Errorf("got: %#v, expected: %#v", got, want)
-			}
-		})
-	}
-}
-
-func TestIsIsolatedRegion(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		input    string
-		expected bool
-	}{
-		{
-			input:    USEast1RegionID,
-			expected: false,
-		},
-		{
-			input:    CNNorth1RegionID,
-			expected: false,
-		},
-		{
-			input:    USGovEast1RegionID,
-			expected: false,
-		},
-		{
-			input:    USISOEast1RegionID,
-			expected: true,
-		},
-		{
-			input:    USISOBEast1RegionID,
-			expected: true,
-		},
-		{
-			input:    EUISOEWest1RegionID,
-			expected: true,
-		},
-	}
-
-	for _, testCase := range testCases {
-		testCase := testCase
-		t.Run(testCase.input, func(t *testing.T) {
-			t.Parallel()
-
-			if got, want := IsIsolatedRegion(testCase.input), testCase.expected; got != want {
-				t.Errorf("got: %#v, expected: %#v", got, want)
-			}
-		})
-	}
-}
-
-func TestIsStandardPartition(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		input    string
-		expected bool
-	}{
-		{
-			input:    StandardPartitionID,
-			expected: true,
-		},
-		{
-			input:    ChinaPartitionID,
-			expected: false,
-		},
-		{
-			input:    USGovCloudPartitionID,
-			expected: false,
-		},
-		{
-			input:    ISOPartitionID,
-			expected: false,
-		},
-		{
-			input:    ISOBPartitionID,
-			expected: false,
-		},
-		{
-			input:    ISOEPartitionID,
-			expected: false,
-		},
-		{
-			input:    ISOFPartitionID,
-			expected: false,
-		},
-	}
-
-	for _, testCase := range testCases {
-		testCase := testCase
-		t.Run(testCase.input, func(t *testing.T) {
-			t.Parallel()
-
-			if got, want := IsStandardPartition(testCase.input), testCase.expected; got != want {
-				t.Errorf("got: %#v, expected: %#v", got, want)
-			}
-		})
-	}
-}
-
-func TestIsStandardRegion(t *testing.T) {
-	t.Parallel()
-
-	testCases := []struct {
-		input    string
-		expected bool
-	}{
-		{
-			input:    USEast1RegionID,
-			expected: true,
-		},
-		{
-			input:    CNNorth1RegionID,
-			expected: false,
-		},
-		{
-			input:    USGovEast1RegionID,
-			expected: false,
-		},
-		{
-			input:    USISOEast1RegionID,
-			expected: false,
-		},
-		{
-			input:    USISOBEast1RegionID,
-			expected: false,
-		},
-		{
-			input:    EUISOEWest1RegionID,
-			expected: false,
-		},
-	}
-
-	for _, testCase := range testCases {
-		testCase := testCase
-		t.Run(testCase.input, func(t *testing.T) {
-			t.Parallel()
-
-			if got, want := IsStandardRegion(testCase.input), testCase.expected; got != want {
-				t.Errorf("got: %#v, expected: %#v", got, want)
-			}
-		})
-	}
-}
